store: split cloudConfig slice into modular file structure
Splits the index.ts into slice.ts (createSlice, actions) and
selectors.ts (state selectors) with a barrel index.ts. No behavioral
changes — all existing import paths resolve unchanged.
